requirements.txt
alembic==1.4.1
appnope==0.1.0
argon2-cffi==20.1.0
asgiref==3.4.1
asn1crypto==0.24.0
async-generator==1.10
attrs==20.2.0
azure-core==1.8.1
azure-storage-blob==12.5.0
backcall==0.2.0
bleach==3.2.0
blis==0.4.1
boto==2.49.0
boto3==1.14.63
botocore==1.17.63
catalogue==1.0.0
catboost==0.24.1
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.4
Click==7.0
cloudpickle==1.6.0
cobble==0.1.3
colorlover==0.3.0
combo==0.1.1
confuse==1.3.0
cryptography==3.4.7
cufflinks==0.17.3
cycler==0.10.0
cymem==2.0.3
databricks-cli==0.11.0
datefinder==0.7.1
decorator==4.4.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.5
django-allauth==0.45.0
django-cors-headers==3.8.0
django-heroku==0.3.1
djangorestframework==3.12.4
docker==4.3.1
docutils==0.15.2
docx==0.2.4
docx2txt==0.8
entrypoints==0.3
Flask==1.1.1
funcy==1.14
future==0.17.1
gensim==3.8.3
gitdb==4.0.5
GitPython==3.1.8
gorilla==0.3.0
graphviz==0.14.1
gunicorn==20.0.4
htmlmin==0.1.12
idna==2.10
ImageHash==4.1.0
imbalanced-learn==0.7.0
importlib-metadata==1.7.0
iniconfig==1.0.1
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.1
kiwisolver==1.2.0
kmodes==0.10.2
lightgbm==3.0.0
llvmlite==0.34.0
lxml==4.7.1
Mako==1.1.3
mammoth==1.4.18
MarkupSafe==1.1.1
matplotlib==3.3.2
missingno==0.4.2
mistune==0.8.4
mlflow==1.11.0
mlxtend==0.17.3
more-itertools==8.5.0
msrest==0.6.19
murmurhash==1.0.2
nbclient==0.5.0
nbconvert==6.0.3
nbformat==5.0.7
nest-asyncio==1.4.0
networkx==2.5
nltk==3.5
notebook==6.1.4
numba==0.51.2
numexpr==2.7.1
numpy==1.19.2
oauthlib==3.1.0
opencv-python==4.5.4.60
packaging==20.4
pandas==1.1.2
pandas-profiling==2.9.0
pandocfilters==1.4.2
parso==0.7.1
patsy==0.5.1
pdflatex==0.1.3
pexpect==4.8.0
phik==0.10.0
pickleshare==0.7.5
Pillow==7.2.0
plac==1.1.3
plotly==4.10.0
pluggy==0.13.1
preshed==3.0.2
projectq==0.7.0
prometheus-client==0.8.0
prometheus-flask-exporter==0.17.0
prompt-toolkit==3.0.7
protobuf==3.13.0
psycopg2==2.7.7
psycopg2-binary==2.9.1
ptyprocess==0.6.0
py==1.9.0
pycaret==2.1.2
pycaret-nightly==2.2.dev1600043546
pycparser==2.20
Pygments==2.7.1
PyJWT==2.1.0
pyLDAvis==2.1.2
pyod==0.8.2
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.0.2
python-dateutil==2.8.1
python-editor==1.0.4
python3-openid==3.2.0
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
querystring-parser==1.2.4
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.0
Send2Trash==1.5.0
six==1.12.0
smart-open==2.1.1
smmap==3.0.4
spacy==2.3.2
SQLAlchemy==1.3.13
sqlparse==0.3.1
srsly==1.0.2
statsmodels==0.12.0
subject==0.1.0
suod==0.0.4
tabulate==0.8.7
tangled-up-in-unicode==0.0.6
terminado==0.8.3
testpath==0.4.4
textblob==0.15.3
thinc==7.4.1
threadpoolctl==2.1.0
toml==0.10.1
tornado==6.0.4
tqdm==4.49.0
traitlets==5.0.4
typing-extensions==3.10.0.0
umap-learn==0.4.6
urllib3==1.25.10
users==1.0.dev0
visions==0.5.0
wasabi==0.8.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==0.16.0
whitenoise==5.3.0
widgetsnbextension==3.5.1
wordcloud==1.8.0
xgboost==1.2.0
yellowbrick==1.1
zipp==3.1.0
